refactor(Registries): - Re-enable tests
